Reverting some more small changes.

+++ src/Camera.cpp
@@ -40,9 +40,9 @@
 #define VIEW_FACTOR 1.50
 
 #define ANGLE(size,dist) \
-	(asin((double)size/dist)*(VIEW_FACTOR*2.0*180.0/M_PI))
+	(asin(size/dist)*(VIEW_FACTOR*2.0*180.0/M_PI))
 #define DIST(size,angle) \
-	(size/sin(angle*(M_PI/(180.0*2.0*VIEW_FACTOR))))
+	(size/(float)sin(angle*(M_PI/(180.0*2.0*VIEW_FACTOR))))
 
 #define MIN_DIST DIST(mainbody->radius,MAX_FOV)
